refactor: merge single-importer modules
