+++ src/tui/board/store.tsx
@@ -12,7 +12,7 @@ import {
 } from "../../domain/task/policy"
 import { getStatus, kagan } from "../../domain/task/metadata"
 import { commandInTaskScope, commandPlan, configuredScopes } from "../../domain/task/commands"
-import { deleteSession } from "../tasks/delete"
+import { deleteSession } from "../tasks"
 import { getFilter, getOrder, setFilter as persistFilter, setOrder } from "../session/preferences"
 import { COLUMNS, type ColumnType } from "../../domain/task/types"
 import { ROUTE, type BoardCard, type BoardSession } from "../types"
